Add testimonials and contact sections; update styles
This commit adds two new sections to the index.html file: "testimonial" and "contact". It also includes new Swiper JavaScript and CSS for the testimonials section. The style.css file has been updated with new styles for the testimonials and contact sections, as well as some adjustments to existing styles. The "setbutton" class has been removed.